% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_with_types.R
\name{collect_with_types}
\alias{collect_with_types}
\title{Force computation of a database query, restoring R types}
\usage{
collect_with_types(tbl)
}
\arguments{
\item{tbl}{An SQL table pointer (generated via \code{\link[dplyr:tbl]{dplyr::tbl()}}).}
}
\description{
collect_with_types() works exactly like \code{\link[dplyr:collect]{dplyr::collect()}} with the
exception that it restores the original R types not saved by the database,
specifically logicals (booleans) and factor levels. It does this by looking
up this information generated from the \code{\link[=dbWriteTable2]{dbWriteTable2()}} function, which
stores a \emph{__types} table in the database for this information.
}
\seealso{
\code{\link[dplyr:collect]{dplyr::collect()}}
}
