% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbwritetable2.R
\name{dbWriteTable2}
\alias{dbWriteTable2}
\title{Copy data frames to database tables, saving R types}
\usage{
dbWriteTable2(con, name, df, ...)
}
\arguments{
\item{con}{A \link[DBI:DBIConnection-class]{DBIConnection} object,
as returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.}

\item{name}{A character string specifying the unqouted DBMS table name.}

\item{df}{a \link[base:data.frame]{data.frame} (or coercible to data.frame).}

\item{...}{Other parameters passed on to methods.}
}
\description{
dbWriteTable2() is a wrapper around \code{\link[DBI:dbWriteTable]{DBI::dbWriteTable()}}; the key exception
is that in addition to saving the relevant database table, it also saves
a special \emph{__types} table which stores R types which SQL databases do not
(necessarily) have, most notably logicals (booleans) in case of sqlite, and
factors more generally. Of course, \emph{__types} becomes a \strong{reserved} table,
and \emph{should not be used for anything else nor changed manually}.
}
\seealso{
\code{\link[DBI:dbWriteTable]{DBI::dbWriteTable()}}
}
