% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose.R
\name{recomposeDF}
\alias{recomposeDF}
\title{Recompose
Recover the original dataframe from the following decomposed representation:
\itemize{
\item a dumbed-down version of the dataframe, where logicals are
coerced to integers, and factors are coerced to strings.
\item a meta dataframe containing the information necessary to
recover the logicals and factors in the original dataframe.
This operation is reversed by \code{\link[=decomposeDF]{decomposeDF()}}.
}}
\usage{
recomposeDF(decomposed)
}
\description{
Recompose
Recover the original dataframe from the following decomposed representation:
\itemize{
\item a dumbed-down version of the dataframe, where logicals are
coerced to integers, and factors are coerced to strings.
\item a meta dataframe containing the information necessary to
recover the logicals and factors in the original dataframe.
This operation is reversed by \code{\link[=decomposeDF]{decomposeDF()}}.
}
}
\examples{
df <- tibble(a = factor(c("a","a","b"), levels = c("a","b"), ordered = TRUE))

decomposeDF(df)

recomposeDF(decomposeDF(df))
}
